Add mechanism for getting component based on configuration
The configuration dictates what sort of component to get and the component returned is the one from the `cop-react-components` library.

In addition to this, the `radios` and `autocomplete` types require options, which necessitates functions for obtaining those.

The logic for retrieving reference data by URL has been mocked within the library for the time being - a proper implementation that uses the URL properly will be needed but that is out of scope for this ticket and PR.