Add FormComponent
Added `FormComponent` to render a form component on a page.

The reason this is in place rather than just using a utility function is the hook for getting hold of the reference data, which must be inside a React component (or inside another hook).

Also added unit tests and some Storybook stories.