Save and use form status
Added a utility function to set up the status of the form, which is added to the payload at the point of saving in a `formStatus` variable.

Correspondingly, when the form initially loads, the renderer will navigate to the appropriate page if it encounters a `data.formStatus` object. Otherwise, it will default to the "first" page, as determined by the form type.