Support containers on form pages
When a form page includes a `Container` component, there is a need to set up the nesting that comes with that. These changes handle that setup.

In addition, removed the utility function for getting editable components on a Page (and its accompanying unit tests), which has been superseded by the utility function for getting editable components within a Container (which also supports nested Containers).