Adjust payload for files
When a file is selected for upload, the Form Renderer now internally adjusts the payload to include a `meta.documents` array with any files that have been selected.

This will allow any consuming application to interrogate the payload to determine when any files need to be uploaded and act appropriately.