Refactor validation
The `FormRenderer` component has been refactored to wrap an `InternalFormRenderer` component inside of a `ValidationContextProvider`, which handles errors and validation across the form now.

This has removed the need for holding onto error states in the form JSON, which made it much harder to track and manage - errors and validation are now accessed within components via a `useValidation()` hook.